rename file to searchTypes
