+++ frontend/src/types/searchTypes.ts
@@ -3,4 +3,5 @@ export interface Opportunity {
   id: number;
   title: string;
   body: string;
+  
 }
